Skiller ut koden for å generere initiell vilkårsvurdering i en egen klasse uten avhengigheter til service/db
